+++ lib/src/options/segmented_button.dart
@@ -51,6 +51,7 @@ class FastSegmentedButton<T> extends FastFormField<Set<T>> {
     super.onChanged,
     super.onReset,
     super.onSaved,
+    super.onTouched,
     super.restorationId,
     super.validator,
     this.emptySelectionAllowed = false,